speer/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
elastic-transport==8.15.1
elasticsearch==8.17.0
PyJWT==2.10.1
pymongo==4.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
